Add auth

+++ SearchApi/Controllers/IndexController.cs
@@ -20,49 +20,62 @@ namespace SearchApi.Controllers
         private readonly UpdateOperations _updateOperations;
         private readonly ObjectCreatorFacade _objectCreatorFacade;
         private readonly ILogger<IndexController> _logger;
+        private readonly UserService _userService;
 
         public IndexController(ObjectCreatorFacade objectCreatorFacade, 
             GetOperations getOperations, 
             UpdateOperations updateOperations, 
-            ILogger<IndexController> logger)
+            ILogger<IndexController> logger, 
+            UserService userService)
         {
             _objectCreatorFacade = objectCreatorFacade;
             _getOperations = getOperations;
             _updateOperations = updateOperations;
             _logger = logger;
+            _userService = userService;
         }
 
         [HttpGet]
-        public async Task<IEnumerable<string>> GetDatabases()
+        public async Task<IActionResult> GetDatabases()
         {
-            return await _getOperations.GetDatabases();
+            if (await _userService.CheckAuthorize(Request) is null)
+                return Unauthorized();
+            return Ok(await _getOperations.GetDatabases());
         }
         
         [HttpGet("index/{dbname}")]
-        public async Task<IEnumerable<string>> GetIndexes(string dbname)
+        public async Task<IActionResult> GetIndexes(string dbname)
         {
-            return await _getOperations.GetIndexes(dbname);
+            if (await _userService.CheckAuthorize(Request, false, dbname) is null)
+                return Unauthorized();
+            return Ok(await _getOperations.GetIndexes(dbname));
         }
         
         [HttpGet("index/{dbname}/{index}/all")]
-        public async Task<IEnumerable<string>> GetDocuments(string dbname, string index)
+        public async Task<IActionResult> GetDocuments(string dbname, string index)
         {
+            if (await _userService.CheckAuthorize(Request, false, dbname) is null)
+                return Unauthorized();
             var docs = await _getOperations.GetDocuments(new IndexModel(dbname, index));
             var result = docs.Select(x => x.ToString());
-            return result;
+            return Ok(result);
         }
         
         [HttpGet("index/{dbname}/{index}/{id}")]
-        public async Task<string> GetDocument(string dbname, string index, string id)
+        public async Task<IActionResult> GetDocument(string dbname, string index, string id)
         {
+            if (await _userService.CheckAuthorize(Request, false, dbname) is null)
+                return Unauthorized();
             var docs = await _getOperations.GetDocuments(new IndexModel(dbname, index));
             var result = docs.FirstOrDefault(x => x.Id.ToString() == id);
-            return result?.ToString();
+            return Ok(result?.ToString());
         }
 
         [HttpPost("index/{dbname}/{index}/add")]
-        public async Task<ActionResult> CreateIndex([FromBody] object obj, string dbname, string index)
+        public async Task<IActionResult> CreateIndex([FromBody] object obj, string dbname, string index)
         {
+            if (await _userService.CheckAuthorize(Request, false, dbname) is null)
+                return Unauthorized();
             try
             {
                 await _objectCreatorFacade.CreateIndexAndAddObject(new IndexModel(dbname, index), obj);
@@ -78,8 +91,10 @@ namespace SearchApi.Controllers
         }
 
         [HttpPut("index/{dbname}/{index}/rename")]
-        public async Task<ActionResult> RenameIndex(string dbname, string index, string name)
+        public async Task<IActionResult> RenameIndex(string dbname, string index, string name)
         {
+            if (await _userService.CheckAuthorize(Request, false, dbname) is null)
+                return Unauthorized();
             try
             {
                 await _updateOperations.RenameIndex(new IndexModel(dbname, index), name);
@@ -95,8 +110,10 @@ namespace SearchApi.Controllers
         }
         
         [HttpPut("index/{dbname}/{index}/{id}/update")]
-        public async Task<ActionResult> UpdateObject([FromBody] object obj,string dbname, string index, string id)
+        public async Task<IActionResult> UpdateObject([FromBody] object obj,string dbname, string index, string id)
         {
+            if (await _userService.CheckAuthorize(Request, false, dbname) is null)
+                return Unauthorized();
             try
             {
                 await _objectCreatorFacade.UpdateObjectAndIndexing(new IndexModel(dbname, index), id, obj);
@@ -112,8 +129,10 @@ namespace SearchApi.Controllers
         }
         
         [HttpDelete("index/{dbname}/delete")]
-        public async Task<ActionResult> DeleteDatabase(string dbname)
+        public async Task<IActionResult> DeleteDatabase(string dbname)
         {
+            if (await _userService.CheckAuthorize(Request, true) is null)
+                return Unauthorized();
             try
             {
                 await DeleteOperations.DeleteDatabase(dbname);
@@ -129,8 +148,10 @@ namespace SearchApi.Controllers
         }
         
         [HttpDelete("index/{dbname}/{index}/delete")]
-        public async Task<ActionResult> DeleteIndex(string dbname, string index)
+        public async Task<IActionResult> DeleteIndex(string dbname, string index)
         {
+            if (await _userService.CheckAuthorize(Request, true) is null)
+                return Unauthorized();
             try
             {
                 await DeleteOperations.DeleteIndex(new IndexModel(dbname, index));
@@ -146,8 +167,10 @@ namespace SearchApi.Controllers
         }
 
         [HttpDelete("index/{dbname}/{index}/{id}/delete")]
-        public async Task<ActionResult> DeleteObject(string dbname, string index, string id)
+        public async Task<IActionResult> DeleteObject(string dbname, string index, string id)
         {
+            if (await _userService.CheckAuthorize(Request, true) is null)
+                return Unauthorized();
             try
             {
                 _logger.Log(LogLevel.Information, $"INFO: Delete object from {dbname}/{index} with id: {id}");
